PCX-2403: remove expired logic from card to session builder
