+++ checkout/src/main/java/com/payoneer/checkout/ui/model/AccountCard.java
@@ -11,10 +11,6 @@ package com.payoneer.checkout.ui.model;
 import static com.payoneer.checkout.ui.model.PaymentSession.LINK_LANGUAGE;
 
 import java.net.URL;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
 import java.util.List;
 import java.util.Map;
 
@@ -76,15 +72,6 @@ public final class AccountCard extends PaymentCard {
     public String getSubtitle() {
         AccountMask accountMask = account.getMaskedAccount();
         return accountMask != null ? PaymentUtils.getExpiryDateString(accountMask) : null;
-    }
-
-    @Override
-    public boolean isExpired() {
-        Instant instantOfNow = Instant.now();
-        AccountMask accountMask = account.getMaskedAccount();
-        LocalDate dateToday
-            = LocalDateTime.ofInstant(instantOfNow, ZoneOffset.systemDefault()).toLocalDate();
-        return accountMask != null && PaymentUtils.isExpired(accountMask, dateToday);
     }
 
     @Override
