+++ checkout/src/main/java/com/payoneer/checkout/ui/service/PaymentSessionBuilder.java
@@ -20,6 +20,10 @@ import static com.payoneer.checkout.model.NetworkOperationType.PRESET;
 import static com.payoneer.checkout.model.NetworkOperationType.UPDATE;
 import static com.payoneer.checkout.model.RegistrationType.NONE;
 
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -28,6 +32,7 @@ import java.util.Map;
 import com.payoneer.checkout.R;
 import com.payoneer.checkout.core.PaymentException;
 import com.payoneer.checkout.localization.LocalizationKey;
+import com.payoneer.checkout.model.AccountMask;
 import com.payoneer.checkout.model.AccountRegistration;
 import com.payoneer.checkout.model.ApplicableNetwork;
 import com.payoneer.checkout.model.ExtraElements;
@@ -148,6 +153,7 @@ public final class PaymentSessionBuilder {
 
         PresetCard card = new PresetCard(account, buttonKey, extraElements);
         card.setCheckable(true);
+        card.setExpired(isExpired(card.getMaskedAccount()));
         return card;
     }
 
@@ -173,6 +179,7 @@ public final class PaymentSessionBuilder {
         card.setDeletable(deletable);
         card.setHideInputForm(!hasFormElements);
         card.setCheckable(true);
+        card.setExpired(isExpired(card.getMaskedAccount()));
 
         int expandedIcon = deletable ? R.drawable.ic_delete : R.drawable.ic_transparent;
         AccountIcon icon = new AccountIcon(R.drawable.ic_edit, expandedIcon);
@@ -183,6 +190,7 @@ public final class PaymentSessionBuilder {
     private AccountCard buildDefaultAccountCard(AccountRegistration account, ListResult listResult) {
         String buttonKey = LocalizationKey.operationButtonKey(account.getOperationType());
         AccountCard card = new AccountCard(account, buttonKey, listResult.getExtraElements());
+        card.setExpired(isExpired(card.getMaskedAccount()));
         boolean deletable = PaymentUtils.toBoolean(listResult.getAllowDelete(), false);
 
         if (deletable) {
@@ -322,5 +330,12 @@ public final class PaymentSessionBuilder {
         if ((card instanceof PresetCard) || cards.size() == 1) {
             card.setPreselected(true);
         }
+    }
+
+    private boolean isExpired(AccountMask accountMask) {
+        Instant instantOfNow = Instant.now();
+        LocalDate dateToday
+            = LocalDateTime.ofInstant(instantOfNow, ZoneOffset.systemDefault()).toLocalDate();
+        return accountMask != null && PaymentUtils.isExpired(accountMask, dateToday);
     }
 }
